+++ pyxelrest/open_api.py
@@ -2,16 +2,17 @@ from collections import OrderedDict
 import datetime
 import logging
 import os
+from pip.commands.install import InstallCommand
 import re
 import requests
 import requests.exceptions
 import sys
 import threading
 import time
-import yaml
 import xlwings
 import xlwings.conversion
 import xlwings.server
+import yaml
 
 try:
     # Python 3
@@ -144,6 +145,19 @@ class ConfigSection:
         if self.api_key:
             self.api_key = convert_environment_variable(self.api_key)
         self.ntlm_auth = service_config.get('ntlm', {})
+        self._install_python_modules(service_config.get('python_modules', []))
+
+    @staticmethod
+    def _install_python_modules(python_modules):
+        try:
+            for extra_module_name in python_modules:
+                result = InstallCommand().main([extra_module_name, '--upgrade', '--disable-pip-version-check'])
+                if result == 0:
+                    logger.info('{0} package updated.'.format(extra_module_name))
+                else:
+                    logger.warning('{0} package update failed.'.format(extra_module_name))
+        except:
+            logger.exception('Unable to install python modules.')
 
     def is_asynchronous(self, udf_return_type):
         return 'async_auto_expand' == udf_return_type